feat(api): add onDraftEditorCopy and onDraftEditorCut for draft-js@0.11
Those methods can be used with the new `onCopy` and `onCut` props introduced in Draft.js 0.11.0,
to replace `registerCopySource` and remove the need for an editor ref. `registerCopySource` is
still available and works the same as before, but will eventually be removed once this package
drops support for Draft.js 0.10.